Move various filestore types in common
