+++ src/main/webapp/ui/src/Export/common.js.flow
@@ -20,3 +20,37 @@ export type ExportSelection =
     |};
 
 export type PageSize = "A4" | "LETTER";
+
+type Path = string;
+
+export type FileLink = {
+  type: "file",
+  size: number,
+  fileSystemFullPath: Path,
+  path: Path,
+};
+
+export type FolderLink = {
+  type: "folder",
+  content: Array<MixedLink>,
+  fileSystemFullPath: Path,
+  size: null,
+};
+
+type FoundFileLink = {|
+  linkType: "file" | "directory",
+  path: Path,
+|};
+export type MixedLink = FileLink | FolderLink;
+
+export type FileSystemId = string;
+
+export type FileSystem = {|
+  id: FileSystemId,
+  name: string,
+  foundNfsLinks: Array<FoundFileLink>,
+  loggedAs: ?string,
+  checkedNfsLinks: Array<?MixedLink>,
+  checkedNfsLinkMessages: { [Path]: string },
+|};
+
