+++ src/main/webapp/ui/src/Export/fileStoreComponents/FoundLinksListing.js
@@ -12,7 +12,7 @@ import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
-import { type FileSystem } from "../ExportFileStore";
+import { type FileSystem } from "../common";
 
 type FoundLinksListingArgs = {|
   // The total number of files found across all of the filestores
